requirements.txt
pygame
discordsdk
pillow